Adds student name and document name to 1up view title bar in Student Workspaces tab
